Add a rudimentary messages scheduler.

+++ backend/src/main/java/com/sim_backend/websockets/OCPPTime.java
@@ -61,6 +61,8 @@ public class OCPPTime implements AutoCloseable {
    */
   @Override
   public void close() throws Exception {
-    this.client.deleteOnReceiveMessage(HeartBeatResponse.class, listener);
+    if (this.client != null) {
+      this.client.deleteOnReceiveMessage(HeartBeatResponse.class, listener);
+    }
   }
 }
